minor restructure
